ProjectFrameworks entity moved to domain.layer
